refactor(decisions): render one browse card in both the grid and the map
The map's list column built its own card, so the two views disagreed:
no owner or revise actions, no revision-requested badge, metrics shown on
drafts, and `canManage` given the value of `isEditable` — which handed a
non-admin author the admin menu.

Both now render `ProposalBrowseCard`, so a proposal looks the same
whichever view you're in. The map keeps only what's actually
view-specific: the hover highlight that syncs with its marker.

Also drops the map's stopPropagation wrapper around the menu. The card
puts `aside` in a `z-10` layer above the title's stretched link, which is
why the grid never needed it.
